feat: 사용하지 않는 UserState 열거형 삭제
- 역할에 따라 파일 위치 변경
