+++ WeatherForecast/WeatherForecast/Controller/LocationManagerDelegate.swift
@@ -53,11 +53,4 @@ final class LocationManagerDelegate: NSObject, CLLocationManagerDelegate {
             }
         }
     }
-    
-    private func toggleUserState(_ manager: CLLocationManager, _ state: UserState) {
-        guard let locationManager = manager as? LocationManager else {
-            print("매니저가 locaionManager가 아니래유")
-            return
-        }
-    }
 }
