+++ WeatherForecast/WeatherForecast/Controller/TableView/EmptyDataSource.swift
@@ -1,0 +1,19 @@
+//
+//  EmptyDataSource.swift
+//  WeatherForecast
+//
+//  Created by Do Yi Lee on 2021/10/20.
+//
+
+import UIKit.UITableView
+
+final class EmptyDataSource: NSObject, UITableViewDataSource {
+    func tableView(_ tableView: UITableView, numberOfRowsInSection section: Int) -> Int {
+        20
+    }
+    
+    func tableView(_ tableView: UITableView, cellForRowAt indexPath: IndexPath) -> UITableViewCell {
+        let cell = tableView.dequeueReusableCell(withIdentifier: FiveDaysForecastCell.identifier, for: indexPath)
+        return cell
+    }
+}
