+++ WeatherForecast/WeatherForecast/Controller/TableView/WeatherTableViewDataSource.swift
@@ -5,7 +5,7 @@
 //  Created by Do Yi Lee on 2021/10/11.
 //
 
-import UIKit
+import UIKit.UITableView
 
 final class WeatherTableviewDataSource: NSObject, UITableViewDataSource {
     func tableView(_ tableView: UITableView, numberOfRowsInSection section: Int) -> Int {
@@ -35,15 +35,4 @@ final class WeatherTableviewDataSource: NSObject, UITableViewDataSource {
 
         return cell
     }    
-}
-
-final class EmptyDataSource: NSObject, UITableViewDataSource {
-    func tableView(_ tableView: UITableView, numberOfRowsInSection section: Int) -> Int {
-        20
-    }
-    
-    func tableView(_ tableView: UITableView, cellForRowAt indexPath: IndexPath) -> UITableViewCell {
-        let cell = tableView.dequeueReusableCell(withIdentifier: FiveDaysForecastCell.identifier, for: indexPath)
-        return cell
-    }
 }
